+++ woof-by-category.php
@@ -2,15 +2,16 @@
 /**
  * Plugin Name: WOOF by Category
  * Plugin URI: https://wordpress.org/plugins/woof-by-category/
- * Description: WooCommerce Product Filter (WOOF) extension to display set of filters depending on current product category page.
+ * Description: WooCommerce Product Filter (WOOF) extension to display set of filters depending on current product
+ * category page.
  * Author: KAGG Design
- * Version: 2.2.1
+ * Version: 2.3
  * Author URI: https://kagg.eu/en/
  * Requires at least: 4.4
- * Tested up to: 5.2
- * Requires PHP: 5.2
+ * Tested up to: 5.3
+ * Requires PHP: 5.6
  * WC requires at least: 3.0
- * WC tested up to: 3.6
+ * WC tested up to: 3.8
  *
  * Text Domain: woof-by-category
  * Domain Path: /languages/
@@ -54,15 +55,10 @@ if ( ! defined( 'WOOF_BY_CATEGORY_VERSION' ) ) {
 /**
  * Init plugin class on plugin load.
  */
+static $woof_by_category_plugin;
 
-static $plugin;
+if ( ! isset( $woof_by_category_plugin ) ) {
+	require_once WOOF_BY_CATEGORY_PATH . '/vendor/autoload.php';
 
-if ( ! isset( $plugin ) ) {
-	if ( version_compare( PHP_VERSION, '5.3.0' ) >= 0 ) {
-		require_once WOOF_BY_CATEGORY_PATH . '/vendor/autoload.php';
-	} else {
-		require_once WOOF_BY_CATEGORY_PATH . '/vendor/autoload_52.php';
-	}
-
-	$plugin = new Woof_By_Category();
+	$woof_by_category_plugin = new Woof_By_Category();
 }